Defer importing `devtools` in `sitecustomize.py`.
It's convenient to be able to use `debug()` as if it were a builtin
function, but importing `devtools` directly in `sitecustomize.py` is a
pretty significant performance hit for every invocation of python that
doesn't use `debug()`.  Specifically, this import takes 100-200 ms,
which is a noticeable delay for interactive sessions and short-running
scripts:

    $ python -X importtime -c 'import devtools'

While it would probably be possible to make `devtools` faster to import,
the best solution is simply to avoid importing `devtools` unless it will
be used.  It took me a while to figure out how to do this, but in the
end the code is pretty clear and succinct, and I think it will be useful
to others.